added Repository and changes in Jpa Annotations

+++ src/main/java/com/nutrition/sweng/Outbound/FoodJpaRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FoodJpaRepository extends CrudRepository<Food, Integer>, FoodRepository {
+public interface FoodJpaRepository extends CrudRepository<Food, Long>, FoodRepository {
 
     List<Food> findByName(String name);
